Corrections for Java version, JAR file added

+++ SBoot_PatManSys/PatientManagementSystem/src/main/java/com/gmck/PatientManagementSystem/UserModel/Services/DoctorService.java
@@ -76,10 +76,10 @@ public class DoctorService implements IUserService {
 	
 	@Override
 	public void deleteUser(String userId) {
-		if(repo.existsById(userId)) {
+		if(repo.existsById(userId) && repo.findAll().size() > 1) {
 			repo.deleteById(userId);
 		}else {
-			ErrorUpdate.getInstance().updateObserver("User doesn't exist");
+			ErrorUpdate.getInstance().updateObserver("There must be one doctor user at all times");
 		}
 	}
